ManagedGroupWithMembers
